requirements.txt
selenium
undetected-chromedriver
openai
python-dotenv
